Refactor process module: replace process export with processTableBook and remove console log
