+++ src/process/index.ts
@@ -1,1 +1,1 @@
-export * from './process';
+export * from './processTableBook';
