+++ src/process/resolveSelector.ts
@@ -35,8 +35,6 @@ export const resolveSelector = (
             column.name
         );
 
-        console.log(column, {pageName, groupName, columnName}, fullname);
-
         if (!columns.has(fullname)) {
             issues.push({ type: 'processing', message: `Invalid column`, path, data: fullname });
             return undefined;
